Various QoL with TSLint and Docs (#12655)
* docs: add README files to all root folders

Except benchmark and tools.

* refactor: move rules/ directory to etc/

And fix a linting error that was found while updating tslint usage

* ci: lint errors out if TSLint shows a warning

This is not a warning from the code itself, but from TSLint. For example,
if a rule was missing, TSLint will just ignore the rule, show something on
STDOUT and not fail.

This fixes CI where local rules were missing for unknown reasons. It will
now fail CI when that happens, which is the correct behaviour.