fix(@ngtools/webpack): support inline style sourcemaps when using css-loader for component styles
When using the `css-loader` as the final entry in the Webpack loader chain, sourcemaps will not
be generated and inlined by the Webpack stylesheet runtime unless the `btoa` global function exists.
Now that the Angular CLI uses the `css-loader` to handle `@import`, the `btoa` function must be
provided within the custom VM execution context used to evaluate the output of the Webpack component
stylesheet processing pipeline.